pyproject.toml
[project]
name = "workdir"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "kantas-spike", email = "101532773+kantas-spike@users.noreply.github.com" }
]
requires-python = ">=3.14"
dependencies = [
    "cookiecutter>=2.6.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
workdir = "workdir:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]
